+++ src/integrationTest/java/io/github/manuelarte/spring/queryparameter/transformers/ClassFieldTransformerImplTest.java
@@ -2,22 +2,31 @@ package io.github.manuelarte.spring.queryparameter.transformers;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import io.github.manuelarte.spring.queryparameter.model.QueryParameterArgumentResolver;
+import io.github.manuelarte.spring.queryparameter.query.QueryCriteria;
+import io.github.manuelarte.spring.queryparameter.transformers.ClassFieldTransformerImplTest.ItConfiguration;
 import java.time.Instant;
 import java.time.LocalDate;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
 @SpringBootTest(classes = ClassFieldTransformerImplTest.class)
 @EnableWebMvc
+@Import({ItConfiguration.class, QueryParameterArgumentResolver.class })
 @ComponentScan("io.github.manuelarte.spring.queryparameter")
 public class ClassFieldTransformerImplTest {
 
   @Autowired
-  private ClassFieldTransformerImpl classFieldTransformer;
+  @Qualifier(("defaultTypeTransformer"))
+  private ClassFieldTransformerImpl<Object> classFieldTransformer;
 
   @Test
   void testStringConversion() {
@@ -75,7 +84,7 @@ public class ClassFieldTransformerImplTest {
   public enum Gender {
     MALE,
     FEMALE,
-    PREFER_NOT_TO_SAY;
+    PREFER_NOT_TO_SAY
   }
 
   public static class ExampleEntity {
@@ -85,6 +94,23 @@ public class ClassFieldTransformerImplTest {
     private LocalDate birthDate;
     private Gender gender;
     private Instant createdAt;
+  }
+
+  @Configuration
+  public static class ItConfiguration {
+    @Bean
+    public QueryCriteriaTransformer<QueryCriteria> queryCriteriaQueryCriteriaTransformer() {
+      return new QueryCriteriaTransformer<QueryCriteria>() {
+        @Override
+        public QueryCriteria apply(final Class<?> aClass, final QueryCriteria queryCriteria) {
+          return queryCriteria;
+        }
+        @Override
+        public boolean supports(Class<?> clazz) {
+          return QueryCriteria.class.equals(clazz);
+        }
+      };
+    }
   }
 
 }