+++ src/main/java/io/github/manuelarte/spring/queryparameter/config/QueryParameterConfig.java
@@ -1,0 +1,12 @@
+package io.github.manuelarte.spring.queryparameter.config;
+
+import io.github.manuelarte.spring.queryparameter.model.TypeTransformerRegistry;
+
+/**
+ * Interface to add custom type transformers.
+ */
+public interface QueryParameterConfig {
+
+  void addTypeTransformer(TypeTransformerRegistry typeTransformerRegistry);
+
+}