+++ src/main/java/io/github/manuelarte/spring/queryparameter/operators/EqualsOperator.java
@@ -1,13 +1,7 @@
 package io.github.manuelarte.spring.queryparameter.operators;
 
-import org.springframework.core.annotation.Order;
-import org.springframework.stereotype.Component;
-
-@Component
-@Order(EqualsOperator.ORDER)
+@lombok.EqualsAndHashCode(callSuper = true)
 public class EqualsOperator extends AbstractMiddleOperator {
-
-  public static final int ORDER = 0;
 
   private static final String OPERATOR = "::";
 
@@ -16,13 +10,4 @@ public class EqualsOperator extends AbstractMiddleOperator {
     return OPERATOR;
   }
 
-  @Override
-  public boolean equals(final Object o) {
-    return this == o || o instanceof EqualsOperator;
-  }
-
-  @Override
-  public int hashCode() {
-    return EqualsOperator.class.hashCode();
-  }
 }