+++ src/main/java/io/github/manuelarte/spring/queryparameter/query/QueryCriterion.java
@@ -1,19 +1,14 @@
 package io.github.manuelarte.spring.queryparameter.query;
 
 import io.github.manuelarte.spring.queryparameter.operators.Operator;
-import java.util.Objects;
 
+@lombok.AllArgsConstructor
+@lombok.EqualsAndHashCode
 public final class QueryCriterion<V> {
 
   private final String key;
   private final Operator<V> operator;
   private final V value;
-
-  public QueryCriterion(final String key, final Operator<V> operator, final V value) {
-    this.key = key;
-    this.operator = operator;
-    this.value = value;
-  }
 
   public static <V> QueryCriterionBuilder<V> builder() {
     return new QueryCriterionBuilder<>();
@@ -29,25 +24,6 @@ public final class QueryCriterion<V> {
 
   public V getValue() {
     return value;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    final QueryCriterion that = (QueryCriterion) o;
-    return Objects.equals(key, that.key)
-        && Objects.equals(operator, that.operator)
-        && Objects.equals(value, that.value);
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(key, operator, value);
   }
 
   @Override