+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1] - 2020-17-04
+## [0.0.2] - 2020-04-23
+
+### Added
+
+- Added query parameter constraints to add in entities NotNull and Null.
+- Added not operator for equals, lower than, lower than or equals, greater than, greater than or equals, in.
+
+## [0.0.1] - 2020-04-17
 
 ### Added